* device-src/s3-device.h (_S3Device): last_byte_read,   next_block_to_read, next_byte_to_read must be gint64
git-svn-id: https://svn.code.sf.net/p/amanda/code/amanda/trunk@7269 a8d146d6-cc15-0410-8900-af154a0219e0

+++ device-src/s3-device.h
@@ -154,9 +154,9 @@ struct _S3Device {
     GThreadPool *thread_pool_read;
     GCond       *thread_idle_cond;
     GMutex      *thread_idle_mutex;
-    int		 last_byte_read;
-    int          next_block_to_read;
-    int		 next_byte_to_read;
+    gint64	 last_byte_read;
+    gint64	 next_block_to_read;
+    gint64	 next_byte_to_read;
     GSList      *objects;
     guint64	 object_size;
     gboolean	 bucket_made;
